chore: restructure btcstaking pkg to align w/ Go
As we are trying to replicate the Golang implementation
of the BTC staking repository, it will be much more easier
to compare those if they follow a similar directory structure.

This PR does the following:
* Aligns the file of the package with the Golang btcstaking library
* Moves the schnorr adaptor signature functionality to its own package
