+++ contracts/btc-staking/src/error.rs
@@ -21,7 +21,7 @@ pub enum ContractError {
     #[error("{0}")]
     Payment(#[from] PaymentError),
     #[error("{0}")]
-    BTCStaking(#[from] babylon_btcstaking::error::Error),
+    BTCStaking(#[from] babylon_btcstaking::errors::Error),
     #[error("error converting from hex to array: {0}")]
     HexArrayError(#[from] HexToArrayError),
     #[error("{0}")]
