+++ packages/btcstaking/README.md
@@ -1,0 +1,5 @@
+# BTC Staking Rust Library
+
+Rust implementation of the BTC Staking library
+implemented in Golang as part of the
+[Babylon Genesis node repository](https://github.com/babylonlabs-io/babylon).
