+++ packages/btcstaking/src/lib.rs
@@ -1,7 +1,5 @@
-pub mod adaptor_sig;
-pub mod error;
-pub mod schnorr;
+pub mod errors;
 pub mod scripts_utils;
-pub mod sig_verify;
-pub mod tx_verify;
-pub type Result<T> = std::result::Result<T, error::Error>;
+pub mod staking;
+pub mod types;
+pub type Result<T> = std::result::Result<T, errors::Error>;
