+++ packages/schnorr-adaptor-signature/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name              = "babylon-schnorr-adaptor-signature"
+description       = "Schnorr signature utilities for Babylon"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+digest          = { workspace = true }
+k256            = { workspace = true }
+sha2            = { workspace = true }
+thiserror       = { workspace = true }
